chore: make component for proposal block
